Fix command/control system to work on actual robot
